add pyth

+++ test/oracle/OracleCustom.t.sol
@@ -13,7 +13,7 @@ contract OracleCustomTest is Test {
     string[] public symbols;
     address[] public oracles;
 
-    function setUp() public virtual { 
+    function setUp() public virtual {
         oc = new OracleCustom();
         oracle = new Oracle();
         symbols.push("XX");
@@ -23,7 +23,7 @@ contract OracleCustomTest is Test {
 
     function test_getPriceByExternal() public {
         console2.log(block.timestamp);
-        (uint p, uint ts) = oracle.getPriceByExternal(address(oc), 1);
+        (uint256 p, uint256 ts) = oracle.getPriceByExternal(address(oc), 1);
         assertEq(p, 0);
         assertEq(ts, 0);
     }
@@ -32,11 +32,10 @@ contract OracleCustomTest is Test {
         vm.warp(2 hours);
         vm.expectRevert("setting price");
         oracle.getPriceByExternal(address(oc), 1);
-        uint price = 2500e18;
+        uint256 price = 2500e18;
         oracle.setFixPrice("XX", 1, price);
-        (uint p, uint ts) = oracle.getPriceByExternal(address(oc), 1);
+        (uint256 p, uint256 ts) = oracle.getPriceByExternal(address(oc), 1);
         assertEq(p, price);
         assertEq(ts, 1);
-
     }
 }
